backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0

mysqlclient>=2.2

django-cors-headers>=4.3

gunicorn>=21.2
pillow>=10.0.0
django-filter
razorpay
python-dotenv>=23.2